// frontend/src/api/generated/models/appSummaryStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Palmimo Portal
 * OpenAPI spec version: 0.1.0
 */

export type AppSummaryStatus = typeof AppSummaryStatus[keyof typeof AppSummaryStatus];


export const AppSummaryStatus = {
  stopped: 'stopped',
  broken: 'broken',
  installing: 'installing',
  updating: 'updating',
  deleting: 'deleting',
  running: 'running',
  starting: 'starting',
  stopping: 'stopping',
  failed: 'failed',
  needs_repair: 'needs_repair',
} as const;